Create Cart.js, Item.js, Order.js and User.js
Cart.js is the model to use for the items in the cart.
Item.js is the model for the items in the store.
Order.js is for the person's items in the order that is finalized. It is similar to Cart.js except for the addition of a date for the order.
User.js is the model used for the person making the order.